Features/sprint01 serhii

+++ FinancialAccounting/Models/NoDBModels/CreateTransactions.cs
@@ -4,7 +4,7 @@
     {
         public long AccountID { get; set; }
         public bool isPositive { get; set; }
-        public long Amount { get; set; }
+        public double Amount { get; set; }
         public long CategoryID { get; set; }
         public string? Discription { get; set; }
     }
@@ -13,7 +13,7 @@
     {
         public long Id { get; set; }
         public bool isPositive { get; set; }
-        public long Amount { get; set; }
+        public double Amount { get; set; }
         public long CategoryID { get; set; }
         public string? Discription { get; set; }
     }
